Put the 'construct' method of detector construction in it's own source file.
